feat: add downloadBlob utility for handling PDF downloads
