+++ tool/frontend/src/utils/pdfUtils.ts
@@ -117,3 +117,18 @@ export const generateRTIPDF = async (data: PDFData): Promise<{ blob: Blob; fileN
 
   return { blob, fileName, finalMarkdown };
 };
+
+/**
+ * Triggers a browser download for a given Blob.
+ * Handles object URL creation and cleanup automatically.
+ */
+export const downloadBlob = (blob: Blob, fileName: string): void => {
+  const url = URL.createObjectURL(blob);
+  const link = document.createElement('a');
+  link.href = url;
+  link.download = fileName;
+  document.body.appendChild(link);
+  link.click();
+  document.body.removeChild(link);
+  URL.revokeObjectURL(url);
+};
